Add files via upload
This program is a change making program. It begins with a stock of 10 quarters, nickels, dimes, and pennies each and then asks the user to enter the price of the purchased item. Then the program outputs the change that needs to be collected and updates the stock until the user tells the program to quit.