Add files via upload
This program asks the user to enter a filename and opens the file. The program then reads the file and categorizes it using lists and loops. The program is designed to find the average salary, median income, salary range, and the cumulative percentage. The program then uses a loop to display the data.