Add files via upload
The program prompts the user to enter a filename and then opens and reads the file. The file contains information regarding crops, states, maximum and minimum amount of the crop grown in that state along with the year that this occurred. The program then builds a list of lists after reading the file. Then the program builds a dictionary from the list of list and finds the minimum and maximum values. Lastly the program uses a loop to display the data.