Add files via upload
This program was developed to play a poker game. The program deals two cards and then deals 5 community cards. The program then goes through every possible function, checking if any of them line up with the cards dealt. The program then checks to see which player wins with the higher cards and outputs the winner and the cards they had. The program then prompts to play again until there are no cards left in the deck to be dealt.